refactor: name sample import workflow by action
